+++ src/observe.js
@@ -28,7 +28,7 @@
         _configuration = {},
 
         // Variables: Attribute Names
-        _attribute_Name_Options = "data-observe-options";
+        _attribute_Name_Watch_Options = "data-observe-watch-options";
 
 
     /*
@@ -57,8 +57,8 @@
     function collectDOMObject( element ) {
         var result = true;
 
-        if ( isDefined( element ) && element.hasAttribute( _attribute_Name_Options ) ) {
-            var bindingOptionsData = element.getAttribute( _attribute_Name_Options );
+        if ( isDefined( element ) && element.hasAttribute( _attribute_Name_Watch_Options ) ) {
+            var bindingOptionsData = element.getAttribute( _attribute_Name_Watch_Options );
 
             if ( isDefinedString( bindingOptionsData ) ) {
                 var bindingOptions = getObjectFromString( bindingOptionsData );
@@ -74,14 +74,14 @@
 
                 } else {
                     if ( !_configuration.safeMode ) {
-                        console.error( "The attribute '" + _attribute_Name_Options + "' is not a valid object." );
+                        console.error( "The attribute '" + _attribute_Name_Watch_Options + "' is not a valid object." );
                         result = false;
                     }
                 }
 
             } else {
                 if ( !_configuration.safeMode ) {
-                    console.error( "The attribute '" + _attribute_Name_Options + "' has not been set correctly." );
+                    console.error( "The attribute '" + _attribute_Name_Watch_Options + "' has not been set correctly." );
                     result = false;
                 }
             }