+++ docs/CHANGE_LOG.md
@@ -5,6 +5,9 @@
 #### **Public Functions:**
 - Added new public function "searchDomForNewWatches()", which is used to search the DOM for new elements to watch, and adds them.
 
+#### **BREAKING CHANGES:**
+- Renamed the DOM attribute "data-observe-options" to "data-observe-watch-options".
+
 <br>